@@ -1,19 +1,38 @@
+import { useEffect, useState } from "react";
 import "./productList.css";
-import { DataGrid } from "@material-ui/data-grid";
-import { DeleteOutline } from "@material-ui/icons";
-import { productRows } from "../../dummyData";
+import { DataGrid } from "@mui/x-data-grid";
+import { DeleteOutline } from "@mui/icons-material";
 import { Link } from "react-router-dom";
-import { useState } from "react";
+import axios from "axios";
+import { useSelector, useDispatch } from "react-redux";
+import { reLoadData, selectLoadData } from "../../redux/loadDataRedux.js";
 
 export default function ProductList() {
-  const [data, setData] = useState(productRows);
+  const [dataProducts, setDataProducts] = useState([]);
+  const dispatch = useDispatch();
+  const loadData = useSelector(selectLoadData);
+  const token = localStorage.getItem("token");
 
-  const handleDelete = (id) => {
-    setData(data.filter((item) => item.id !== id));
+  useEffect(() => {
+    const getProducts = async () => {
+      const { data } = await axios.get("/products");
+      setDataProducts(data);
+    };
+    getProducts();
+  }, [loadData]);
+
+  const handleDelete = async (id) => {
+    const res = await axios.delete(`/products/${id}`, {
+      headers: { token: `Bearer ${token}` },
+    });
+    console.log(res);
+    if (res.status === 200) {
+      dispatch(reLoadData());
+    }
   };
 
   const columns = [
-    { field: "id", headerName: "ID", width: 90 },
+    { field: "_id", headerName: "ID", width: 300 },
     {
       field: "product",
       headerName: "Product",
@@ -21,36 +40,30 @@ export default function ProductList() {
       renderCell: (params) => {
         return (
           <div className="productListItem">
-            <img className="productListImg" src={params.row.img} alt="" />
+            <img className="productListImg" src={params.row.image} alt="" />
             {params.row.name}
           </div>
         );
       },
     },
-    { field: "stock", headerName: "Stock", width: 200 },
-    {
-      field: "status",
-      headerName: "Status",
-      width: 120,
-    },
     {
       field: "price",
       headerName: "Price",
-      width: 160,
+      width: 300,
     },
     {
       field: "action",
       headerName: "Action",
-      width: 150,
+      width: 300,
       renderCell: (params) => {
         return (
           <>
-            <Link to={"/product/" + params.row.id}>
+            <Link to={"/sanpham/" + params.row._id} state={params.row._id}>
               <button className="productListEdit">Edit</button>
             </Link>
             <DeleteOutline
               className="productListDelete"
-              onClick={() => handleDelete(params.row.id)}
+              onClick={() => handleDelete(params.row._id)}
             />
           </>
         );
@@ -61,9 +74,10 @@ export default function ProductList() {
   return (
     <div className="productList">
       <DataGrid
-        rows={data}
+        rows={dataProducts}
         disableSelectionOnClick
         columns={columns}
+        getRowId={(row) => row._id}
         pageSize={8}
         checkboxSelection
       />
