@@ -1,123 +1,151 @@
+import "./user.css";
 import {
   CalendarToday,
+  LineAxisOutlined,
   LocationSearching,
   MailOutline,
   PermIdentity,
   PhoneAndroid,
   Publish,
-} from "@material-ui/icons";
-import { Link } from "react-router-dom";
-import "./user.css";
+} from "@mui/icons-material";
+import { useEffect, useState } from "react";
+import { Link, useLocation } from "react-router-dom";
+import axios from "axios";
+import { useSelector, useDispatch } from "react-redux";
+import { reLoadData, selectLoadData } from "../../redux/loadDataRedux.js";
+import { Avatar } from "@mui/material";
+import { ToastContainer, toast } from "react-toastify";
+import "react-toastify/dist/ReactToastify.css";
 
 export default function User() {
+  const [infoUser, setInfoUser] = useState([]);
+  const [updateData, setUpdateData] = useState({
+    username: "",
+    email: "",
+  });
+  const loadData = useSelector(selectLoadData);
+  const dispatch = useDispatch();
+  const userId = useLocation().state;
+  const token = localStorage.getItem("token");
+
+  const toastOptions = {
+    position: "top-right",
+    autoClose: 3000,
+    pauseOnHover: true,
+    draggable: true,
+    theme: "dark",
+  };
+
+  useEffect(() => {
+    const getInfoUser = async () => {
+      const { data } = await axios.get(`/users/find/${userId}`, {
+        headers: { token: `Bearer ${token}` },
+      });
+      setInfoUser(data);
+    };
+    getInfoUser();
+  }, [loadData]);
+
+  const changeValue = (e) => {
+    setUpdateData((prev) => ({ ...prev, [e.target.name]: e.target.value }));
+  };
+
+  const handleUpdate = async (e) => {
+    e.preventDefault();
+    const { username, email } = updateData;
+    const res = await axios.put(
+      `/users/${userId}`,
+      {
+        username,
+        email,
+      },
+      {
+        headers: { token: `Bearer ${token}` },
+      }
+    );
+    if (res.status === 200) {
+      dispatch(reLoadData());
+      setUpdateData({ username: "", email: "" });
+      toast.success("Update Thành Công", toastOptions);
+    }
+  };
+
   return (
-    <div className="user">
-      <div className="userTitleContainer">
-        <h1 className="userTitle">Edit User</h1>
-        <Link to="/newUser">
-          <button className="userAddButton">Create</button>
-        </Link>
-      </div>
-      <div className="userContainer">
-        <div className="userShow">
-          <div className="userShowTop">
-            <img
-              src="https://images.pexels.com/photos/1152994/pexels-photo-1152994.jpeg?auto=compress&cs=tinysrgb&dpr=2&w=500"
-              alt=""
-              className="userShowImg"
-            />
-            <div className="userShowTopTitle">
-              <span className="userShowUsername">Anna Becker</span>
-              <span className="userShowUserTitle">Software Engineer</span>
+    <>
+      <div className="user">
+        <div className="userTitleContainer">
+          <h1 className="userTitle">Edit User</h1>
+          <button className="BtnUserUpdate" onClick={handleUpdate}>
+            Update
+          </button>
+        </div>
+        <div className="userContainer">
+          <div className="userShow">
+            <div className="userShowTop">
+              <Avatar />
+              <div className="userShowTopTitle">
+                <span className="userShowUsername">{infoUser.username}</span>
+                <span className="userShowUserTitle">Software Engineer</span>
+              </div>
+            </div>
+            <div className="userShowBottom">
+              <span className="userShowTitle">Account Details</span>
+              <div className="userShowInfo">
+                <PermIdentity className="userShowIcon" />
+                <span className="userShowInfoTitle">{infoUser.username}</span>
+              </div>
+              <div className="userShowInfo">
+                <CalendarToday className="userShowIcon" />
+                <span className="userShowInfoTitle">01.01.1999</span>
+              </div>
+              <span className="userShowTitle">Contact Details</span>
+              <div className="userShowInfo">
+                <PhoneAndroid className="userShowIcon" />
+                <span className="userShowInfoTitle">0123456789</span>
+              </div>
+              <div className="userShowInfo">
+                <MailOutline className="userShowIcon" />
+                <span className="userShowInfoTitle">{infoUser.email}</span>
+              </div>
+              <div className="userShowInfo">
+                <LocationSearching className="userShowIcon" />
+                <span className="userShowInfoTitle">Việt Nam</span>
+              </div>
             </div>
           </div>
-          <div className="userShowBottom">
-            <span className="userShowTitle">Account Details</span>
-            <div className="userShowInfo">
-              <PermIdentity className="userShowIcon" />
-              <span className="userShowInfoTitle">annabeck99</span>
-            </div>
-            <div className="userShowInfo">
-              <CalendarToday className="userShowIcon" />
-              <span className="userShowInfoTitle">10.12.1999</span>
-            </div>
-            <span className="userShowTitle">Contact Details</span>
-            <div className="userShowInfo">
-              <PhoneAndroid className="userShowIcon" />
-              <span className="userShowInfoTitle">+1 123 456 67</span>
-            </div>
-            <div className="userShowInfo">
-              <MailOutline className="userShowIcon" />
-              <span className="userShowInfoTitle">annabeck99@gmail.com</span>
-            </div>
-            <div className="userShowInfo">
-              <LocationSearching className="userShowIcon" />
-              <span className="userShowInfoTitle">New York | USA</span>
-            </div>
+          <div className="userUpdate">
+            <span className="userUpdateTitle">Edit</span>
+            <form className="userUpdateForm">
+              <div className="userUpdateLeft">
+                <div className="userUpdateItem">
+                  <label>Tên Đăng Nhập</label>
+                  <input
+                    type="text"
+                    name="username"
+                    value={updateData.username}
+                    placeholder={infoUser.username}
+                    className="userUpdateInput"
+                    onChange={changeValue}
+                  />
+                </div>
+
+                <div className="userUpdateItem">
+                  <label>Email</label>
+                  <input
+                    type="text"
+                    name="email"
+                    value={updateData.email}
+                    placeholder={infoUser.email}
+                    className="userUpdateInput"
+                    onChange={changeValue}
+                  />
+                </div>
+              </div>
+            </form>
           </div>
         </div>
-        <div className="userUpdate">
-          <span className="userUpdateTitle">Edit</span>
-          <form className="userUpdateForm">
-            <div className="userUpdateLeft">
-              <div className="userUpdateItem">
-                <label>Username</label>
-                <input
-                  type="text"
-                  placeholder="annabeck99"
-                  className="userUpdateInput"
-                />
-              </div>
-              <div className="userUpdateItem">
-                <label>Full Name</label>
-                <input
-                  type="text"
-                  placeholder="Anna Becker"
-                  className="userUpdateInput"
-                />
-              </div>
-              <div className="userUpdateItem">
-                <label>Email</label>
-                <input
-                  type="text"
-                  placeholder="annabeck99@gmail.com"
-                  className="userUpdateInput"
-                />
-              </div>
-              <div className="userUpdateItem">
-                <label>Phone</label>
-                <input
-                  type="text"
-                  placeholder="+1 123 456 67"
-                  className="userUpdateInput"
-                />
-              </div>
-              <div className="userUpdateItem">
-                <label>Address</label>
-                <input
-                  type="text"
-                  placeholder="New York | USA"
-                  className="userUpdateInput"
-                />
-              </div>
-            </div>
-            <div className="userUpdateRight">
-              <div className="userUpdateUpload">
-                <img
-                  className="userUpdateImg"
-                  src="https://images.pexels.com/photos/1152994/pexels-photo-1152994.jpeg?auto=compress&cs=tinysrgb&dpr=2&w=500"
-                  alt=""
-                />
-                <label htmlFor="file">
-                  <Publish className="userUpdateIcon" />
-                </label>
-                <input type="file" id="file" style={{ display: "none" }} />
-              </div>
-              <button className="userUpdateButton">Update</button>
-            </div>
-          </form>
-        </div>
       </div>
-    </div>
+      <ToastContainer />
+    </>
   );
 }
