@@ -1,56 +1,69 @@
 import "./userList.css";
-import { DataGrid } from "@material-ui/data-grid";
-import { DeleteOutline } from "@material-ui/icons";
-import { userRows } from "../../dummyData";
+import { DataGrid } from "@mui/x-data-grid";
+import { DeleteOutline } from "@mui/icons-material";
 import { Link } from "react-router-dom";
-import { useState } from "react";
+import { useEffect, useState } from "react";
+import axios from "axios";
+import { useSelector, useDispatch } from "react-redux";
+import { reLoadData, selectLoadData } from "../../redux/loadDataRedux.js";
+import { Avatar } from "@mui/material";
 
 export default function UserList() {
-  const [data, setData] = useState(userRows);
+  const [dataUser, setDataUser] = useState([]);
+  const dispatch = useDispatch();
+  const loadData = useSelector(selectLoadData);
+  const token = localStorage.getItem("token");
 
-  const handleDelete = (id) => {
-    setData(data.filter((item) => item.id !== id));
+  useEffect(() => {
+    const getUser = async () => {
+      const { data } = await axios.get("/users", {
+        headers: { token: `Bearer ${token}` },
+      });
+      setDataUser(data);
+    };
+    getUser();
+  }, [loadData]);
+
+  const handleDelete = async (id) => {
+    const res = await axios.delete(`/users/${id}`, {
+      headers: { token: `Bearer ${token}` },
+    });
+    console.log(res);
+    if (res.status === 200) {
+      dispatch(reLoadData());
+    }
   };
-  
+
   const columns = [
-    { field: "id", headerName: "ID", width: 90 },
+    { field: "_id", headerName: "ID", width: 300 },
     {
       field: "user",
       headerName: "User",
-      width: 200,
+      width: 300,
       renderCell: (params) => {
         return (
           <div className="userListUser">
-            <img className="userListImg" src={params.row.avatar} alt="" />
+            <Avatar />
             {params.row.username}
           </div>
         );
       },
     },
-    { field: "email", headerName: "Email", width: 200 },
-    {
-      field: "status",
-      headerName: "Status",
-      width: 120,
-    },
-    {
-      field: "transaction",
-      headerName: "Transaction Volume",
-      width: 160,
-    },
+    { field: "email", headerName: "Email", width: 300 },
+
     {
       field: "action",
       headerName: "Action",
-      width: 150,
+      width: 300,
       renderCell: (params) => {
         return (
           <>
-            <Link to={"/user/" + params.row.id}>
+            <Link to={"/nguoidung/" + params.row._id} state={params.row._id}>
               <button className="userListEdit">Edit</button>
             </Link>
             <DeleteOutline
               className="userListDelete"
-              onClick={() => handleDelete(params.row.id)}
+              onClick={() => handleDelete(params.row._id)}
             />
           </>
         );
@@ -61,9 +74,10 @@ export default function UserList() {
   return (
     <div className="userList">
       <DataGrid
-        rows={data}
+        rows={dataUser}
         disableSelectionOnClick
         columns={columns}
+        getRowId={(row) => row._id}
         pageSize={8}
         checkboxSelection
       />
