@@ -14,7 +14,7 @@ router.put("/:id", verifyToken, UpdateUser);
 //DELETE
 router.delete("/:id", verifyToken, DeleteUser);
 //GET USER
-router.get("/find/:id", getUser);
+router.get("/find/:id", verifyToken, getUser);
 //GET ALL USER
 router.get("/", verifyToken, getAllUser);
 router.get("/stats", verifyToken, getUserStats);
